Split the test file in multiple files to match the module structure
